Fix a real cipher-naming bug, add hover formulas, ship connections + notable values
The "exotic" cipher I'd been calling Squared Value (Meruba) was only ever
computing Mispar P'rati (sum of each letter's value squared). A distinct,
equally real traditional system, Mispar HaMerubah HaKlali (the word's
total value squared *once*), also goes by "squared" and produces a much
larger, genuinely different number -- for longer words the gap between
the two grows enormously, which is almost certainly what looked "clearly
wrong at higher values." Both are now implemented and clearly
distinguished (P'rati and Klali), verified against Wikipedia's gematria
reference rather than memory. Also switched the per-cipher index from
Int32Array to Float64Array, since Klali's squaring can exceed the 32-bit
range for long search input.

Every cipher badge now has a hover tooltip spelling out its exact
computation, sourced from the same formula descriptions in gematria.js.

Improved English-translation highlight coverage (44%->46% of words,
15%->15.5% of phrases) by trying a few regular English inflections
(plural/singular, simple past) of the candidate gloss word before giving
up -- the gap the "not uniform" feedback was pointing at.

Stage 1 (Connections): each result now shows when it also matches the
search under other systems, computed directly from the same precomputed
per-cipher values already in the index -- with the two pairs that are
mathematically forced to match together (Total Squared and Integral
Reduced both derive deterministically from Standard Value) explicitly
excluded, since flagging a guaranteed match as a "coincidence" would be
worse than not flagging it at all.

Stage 2 (Notable values): a small, hand-verified glossary of well-known
gematria values (src/notable-values.js) shown as a callout when a search
lands on one -- every entry's arithmetic is checked against the app's own
cipher functions, not just asserted.

+++ src/data.js
@@ -48,13 +48,16 @@ async function fetchJson(path) {
 }
 
 /**
- * Builds one Int32Array per cipher holding the precomputed value for every
+ * Builds one Float64Array per cipher holding the precomputed value for every
  * item (parallel to `items`), so a match query is a single linear scan.
+ * Float64 (not Int32) because systems like Total Squared (klali) can exceed
+ * the 32-bit range for long search input, and JS numbers stay exact up to
+ * 2^53 either way.
  */
 function buildIndex(items, getText) {
   const n = items.length;
   const values = {};
-  for (const key of HE_CIPHER_KEYS) values[key] = new Int32Array(n);
+  for (const key of HE_CIPHER_KEYS) values[key] = new Float64Array(n);
 
   for (let i = 0; i < n; i++) {
     const vec = computeHebrewCipherVector(getText(items[i]));
